WIP: Интерпретация несложной программы на Swift

+++ interpreter/ast/AbstractSyntaxTree.java
@@ -1,6 +1,6 @@
 package interpreter.ast;
 
 public interface AbstractSyntaxTree {
-    int calculate();
+    double calculate();
 }
