+++ interpreter/ast/Compound.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record Compound(List<AbstractSyntaxTree> statementList) implements AbstractSyntaxTree {
 
     @Override
-    public int calculate() {
+    public double calculate() {
         for (AbstractSyntaxTree statement : statementList) {
             statement.calculate();
         }
